fitria-plan/requirements.txt
streamlit
pandas
python-dotenv
gspread
google-auth
icalendar
pytz
openpyxl